requirement.txt
videosdk-python>=1.2.1  
whispercpp>=0.0.14     
gtts>=2.4.0           
aiohttp>=3.8.0          
transformers>=4.30.0    
langchain>=0.0.200      
faiss-cpu>=1.7.4        
rank-bm25>=0.2.2        
watchdog>=3.0.0         
google-api-python-client>=2.80.0  
tenacity>=8.2.0        
pydantic>=1.10.0        
numpy>=1.24.0          
webrtcvad>=2.3.5       